beta test 10

+++ uninstall.sh
@@ -16,14 +16,21 @@ SYSTEMD=/etc/systemd/system
 SYSTEMCTL=$(which systemctl)
 
 if [ `id -u` != "0" ]; then
-    echo "Error at uinstaller, please run uinstaller as root user"
+    echo "Error at uninstallation, please run uninstaller as root"
     exit 1
 fi
 
 echo "Uninstalling NoobzVpn-Server..."
-$SYSTEMCTL stop noobzvpns.service
-$SYSTEMCTL disable noobzvpns.service
-rm $SYSTEMD/noobzvpns.service
-rm $BIN/noobzvpns
-rm -rf $CONFIGS
+if [ -f $SYSTEMD/noobzvpns.service ]; then
+    $SYSTEMCTL stop noobzvpns.service
+    $SYSTEMCTL disable noobzvpns.service
+    rm $SYSTEMD/noobzvpns.service
+    $SYSTEMCTL daemon-reload
+fi
+if [ -f $BIN/noobzvpns ]; then
+    rm $BIN/noobzvpns
+fi
+if [ -d $CONFIGS ]; then
+    rm -rf $CONFIGS
+fi
 echo "Uninstall NoobzVpn-Server completed"
